+++ HelloHome.Central.Hub/MessageChannel/SerialPortMessageChannel/Encoders/NodeConfigEncoder.cs
@@ -18,10 +18,10 @@ namespace HelloHome.Central.Hub.MessageChannel.SerialPortMessageChannel.Encoders
 		protected override void EncodeBody (NodeConfigCommand message, List<byte> bytes)
 		{
 			bytes.AddRange(BitConverter.GetBytes((long)message.Signature));
-			bytes.AddRange(BitConverter.GetBytes(message.NewRfAddress));
-			bytes.AddRange(BitConverter.GetBytes((short)message.ExtraFeatures));
-		    bytes.Add(message.NodeInfoFreq);
-		    bytes.Add(message.EnvironmentFreq);
+			bytes.AddRange(BitConverter.GetBytes((UInt16)message.NewRfAddress));
+			bytes.AddRange(BitConverter.GetBytes((UInt16)message.ExtraFeatures));
+		    bytes.Add((byte)message.NodeInfoFreq);
+		    bytes.Add((byte)message.EnvironmentFreq);
 		}
 		protected override byte Discriminator => Constants.Message.Command.NodeConfigCommand;
 	}
